Safari-611.1.10/LayoutTests/fullscreen/fullscreen-restore-scroll-position-expected.txt

This tests that page scroll is restored after fullscreen. Press any key to start the test.


EVENT(load)
EXPECTED ((document.scrollingElement.scrollTop === originalScroll) == 'true') OK
